[AN] feat(#505): 스터디 가입 신청 리스트 UI 구현
