+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/mystudy/StudyApplyMemberProfileFragment.kt
@@ -1,0 +1,23 @@
+package com.takseha.presentation.ui.mystudy
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.takseha.presentation.R
+
+class StudyApplyMemberProfileFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_study_apply_member_profile, container, false)
+    }
+}
