+++ android/gitudy/presentation/src/main/res/layout/activity_main_home_alert.xml
@@ -6,86 +6,12 @@
     android:layout_height="match_parent"
     tools:context=".ui.home.MainHomeAlertActivity">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/constraintLayout4"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/mainHomeAlertContainer"
+        android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:paddingLeft="22dp"
-        android:paddingRight="22dp"
-        android:paddingBottom="47dp">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/alertTitle"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="20dp"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <ImageView
-                android:id="@+id/backBtn"
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:src="@drawable/ic_arrow_back"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:id="@+id/textView2"
-                style="@style/B3_1"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="알림"
-                android:textColor="@color/BLACK"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <ImageView
-                android:id="@+id/deleteAllBtn"
-                android:layout_width="35dp"
-                android:layout_height="35dp"
-                android:src="@drawable/ic_delete_all"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-        <LinearLayout
-            android:id="@+id/isNoAlertLayout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:gravity="center"
-            android:orientation="vertical"
-            android:visibility="gone"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <ImageView
-                android:layout_width="40dp"
-                android:layout_height="40dp"
-                android:layout_marginBottom="5dp"
-                android:src="@drawable/ic_alert_300" />
-
-            <TextView
-                style="@style/B3_2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:gravity="center"
-                android:text="@string/notice_no_alert"
-                android:textColor="@color/GS_400" />
-        </LinearLayout>
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/alertList"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="30dp"
-            app:layout_constraintTop_toBottomOf="@+id/alertTitle"
-            tools:layout_editor_absoluteX="22dp" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
+        app:defaultNavHost="true"
+        app:layout_constraintStart_toStartOf="parent"
+        app:navGraph="@navigation/main_home_alert_nav" />
 </androidx.constraintlayout.widget.ConstraintLayout>
